src/macport/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
ply==3.11
plyara==2.1.1
pytest==7.3.0
python-markdown-generator @ git+https://github.com/Nicceboy/python-markdown-generator@8b8801f39a34d053f53e3a2eb76a8553e6bee7eb
requests==2.28.2
simple-ddl-parser==0.30.0
sql-metadata==2.7.0
sqlparse==0.4.3
urllib3==1.26.15
